chore: rearrange code appropriately
